refactor: add jsdocs and clean up imports
