+++ src/app/match/[matchId]/components/Scoreboard/PlayerScoreboard.tsx
@@ -1,5 +1,5 @@
-import { CSGoldVision } from '@/app/match/[matchId]/components/Scoreboard/CSGoldVision';
-import { DamageMeter } from '@/app/match/[matchId]/components/Scoreboard/DamageMeter';
+import { CSGoldVision } from './CSGoldVision';
+import { DamageMeter } from './DamageMeter';
 import { ChampIcon } from '@/components/riotIcons/ChampIcon';
 import { ChampScoreboard } from '@/components/riotIcons/ChampScoreboard';
 import { Item } from '@/components/riotIcons/Item';
@@ -8,17 +8,27 @@ import { Rune } from '@/components/riotIcons/Rune';
 import { SummonerSpell } from '@/components/riotIcons/SummonerSpell';
 import { cn } from '@/lib/utils';
 
+/** Props for {@link PlayerScoreboard}. */
 export type PlayerScoreboardProps = {
+  /** The Riot post-match participant data. */
   participant: Riot.MatchV5.Participant;
+  /** The most total damage dealt to champions by any player in the match. */
   maxDamage: number;
+  /** The most damage taken from champions by any player in the match. */
   maxDamageTaken: number;
+  /** The highest CC-score of any player in the match. */
   maxCC: number;
+  /** The size of the scoreboard component. */
   size?: 'sm' | 'md' | 'lg';
+  /** Whether to use the champion icon layout instead of the splash layout. */
   icon?: boolean;
+  /** The scoreboard's controller key. */
   group?: string;
+  /** The patch the match is played on, using latest assets if undefined. */
   patch?: string;
 };
 
+/** An individual player's row in the match scoreboard. */
 export const PlayerScoreboard = ({
   participant,
   maxDamage,
